# swagger.yaml
swagger: '2.0'
info:
  title: Fundr API Documentation
  version: 1.0.0
  description: API documentation for the Fundr application
basePath: /api
schemes:
  - http
  - https
paths:
  /users:
    post:
      summary: Create a new user
      tags:
        - User
      parameters:
        - name: user
          in: body
          description: User data
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: User created successfully
        '500':
          description: Internal server error
    get:
      summary: Get all users
      tags:
        - User
      responses:
        '200':
          description: List of users
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '500':
          description: Internal server error

  /users/{id}:
    get:
      summary: Get a user by ID
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          type: integer
      responses:
        '200':
          description: User found
          schema:
            $ref: '#/definitions/User'
        '404':
          description: User not found
        '500':
          description: Internal server error
    put:
      summary: Update a user by ID
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          type: integer
        - name: user
          in: body
          description: Updated user data
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: User updated successfully
        '404':
          description: User not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a user by ID
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          type: integer
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
        '500':
          description: Internal server error

  /campaigns:
    post:
      summary: Create a new fundraising campaign
      tags:
        - FundraisingCampaign
      parameters:
        - name: campaign
          in: body
          description: Campaign data
          required: true
          schema:
            $ref: '#/definitions/FundraisingCampaign'
      responses:
        '201':
          description: Campaign created successfully
        '500':
          description: Internal server error
    get:
      summary: Get all fundraising campaigns
      tags:
        - FundraisingCampaign
      responses:
        '200':
          description: List of campaigns
          schema:
            type: array
            items:
              $ref: '#/definitions/FundraisingCampaign'
        '500':
          description: Internal server error

  /campaigns/{id}:
    get:
      summary: Get a campaign by ID
      tags:
        - FundraisingCampaign
      parameters:
        - name: id
          in: path
          description: Campaign ID
          required: true
          type: integer
      responses:
        '200':
          description: Campaign found
          schema:
            $ref: '#/definitions/FundraisingCampaign'
        '404':
          description: Campaign not found
        '500':
          description: Internal server error
    put:
      summary: Update a campaign by ID
      tags:
        - FundraisingCampaign
      parameters:
        - name: id
          in: path
          description: Campaign ID
          required: true
          type: integer
        - name: campaign
          in: body
          description: Updated campaign data
          required: true
          schema:
            $ref: '#/definitions/FundraisingCampaign'
      responses:
        '200':
          description: Campaign updated successfully
        '404':
          description: Campaign not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a campaign by ID
      tags:
        - FundraisingCampaign
      parameters:
        - name: id
          in: path
          description: Campaign ID
          required: true
          type: integer
      responses:
        '204':
          description: Campaign deleted successfully
        '404':
          description: Campaign not found
        '500':
          description: Internal server error

definitions:
  User:
    type: object
    properties:
      email:
        type: string
      username:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      phone_number:
        type: string
  FundraisingCampaign:
    type: object
    properties:
      name:
        type: string
      duration:
        type: integer
      targetedAmount:
        type: number
      documentProof:
        type: string
      description:
        type: string
      hospitalAccountDetails:
        type: string
      creatorId:
        type: integer